Refactor the code structure of test utils
